backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
selenium
webdriver-manager
python-dotenv
selenium-stealth